+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## 1.0.0 - 2024-04-06
 
 ### Added
-- Initial release
+- Initial release for the Bunny => BunnyMate rebrand
+- Added the ability to pass assets directly to the `bunnyPullUrl()` Twig function